fastApi/app/requirements.txt
fastapi==0.95.1
starlette==0.26.1
numpy==1.23.5
pillow==10.0.1
SQLAlchemy==2.0.22
SQLAlchemy-Utils==0.41.1
urllib3==1.26.17
uvicorn==0.22.0
requests==2.28.2
python-multipart==0.0.6
typing-extensions==4.7.1
celery==5.3.6
redis==5.0.1
bcrypt==4.1.2
PyJWT==2.8.0
email-validator==2.1.0.post1

